api/requirements.txt
appnope==0.1.0
argh==0.26.2
asn1crypto==0.24.0
attrs==17.4.0
bandit==1.4.0
bcrypt==3.1.4
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
coverage==4.5.1
cryptography==2.1.4
decorator==4.2.1
dj-database-url==0.5.0
Django==2.0.2
django-extensions==2.0.0
djangorestframework==3.7.7
djangorestframework-jsonapi==2.4.0
docopt==0.6.2
enum34==1.1.6
facebook-sdk==2.0.0
factory-boy==2.10.0
Faker==0.8.11
flake8==3.5.0
flake8-author==1.1.4
flake8-bandit==1.0.1
flake8-bugbear==18.2.0
flake8-commas==1.0.0
flake8-docstrings==1.3.0
flake8-import-order==0.17
flake8-junit-report==2.1.0
flake8-module-name==0.1.5
flake8-polyfill==1.0.2
flake8-quotes==0.14.0
gitdb2==2.0.3
GitPython==2.1.8
google-api-python-client==1.6.5
hacking==1.0.0
httplib2==0.10.3
idna==2.6
inflection==0.3.1
invoke==0.22.1
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
mccabe==0.6.1
mock==2.0.0
oauth2client==4.1.2
paramiko==2.4.0
parso==0.1.1
pathtools==0.1.2
pbr==3.1.1
pep8==1.7.1
pep8-naming==0.5.0
pexpect==4.4.0
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2==2.7.3
ptyprocess==0.5.2
py==1.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.3.1
pycparser==2.18
pydocstyle==2.1.1
pyflakes==1.6.0
Pygments==2.2.0
PyJWT==1.6.0
PyNaCl==1.2.1
pytest==3.4.1
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-flake8==0.9.1
pytest-mock==1.7.1
pytest-sugar==0.9.1
pytest-watch==4.1.0
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
simplegeneric==0.8.1
six==1.11.0
smmap2==2.0.3
snowballstemmer==1.2.1
stevedore==1.28.0
termcolor==1.1.0
text-unidecode==1.2
traitlets==4.3.2
typing==3.6.4
uritemplate==3.0.0
urllib3==1.22
watchdog==0.8.3
wcwidth==0.1.7
